+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), ~~and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)~~.
 
 ## [Unreleased]
+### Added
+ - Add document language filter ([#20](https://github.com/pwyf/iati-decipher/issues/20))
+
 ### Fixed
  - Fix bug with x-axis tick values and missing legend values ([#100](https://github.com/pwyf/iati-decipher/issues/100))